agregaod userrole

+++ app/controllers/DoctorController.php
@@ -11,7 +11,12 @@ class DoctorController extends SecretaryController {
   }
   public function doUnassignSecretary()
   {
+    
     SecretaryDoctor::find(Input::get('idSecretaryDoctor'))->delete();
+    
+    /*UserRole::where('rut', '=', $secDoc->secretarysRut)
+      ->where('idHospital', '=', $secDoc->idHospital)
+      ->where('role', '=', 'secretary');*/
   }
   
   public function doAssignSecretary()
@@ -28,32 +33,51 @@ class DoctorController extends SecretaryController {
     }
     $rutSinPuntos = str_replace(".", "", Input::get('rut'));
     list($rut, $dv) = explode("-", $rutSinPuntos);
-    $secretary = User::where('role', '=', 'secretary')
-      ->where('rut', '=', $rut)->first();
+    $secretary = User::where('rut', '=', $rut)->first();
+
     
     if (!isset($secretary))
     {
-      $secretary = new User();
-      $secretaryInfo = new UserInfo();
       
+      $secretary = new User();
+      
+      $secretary->name = null;
+      $secretary->lastname = null;
       $secretary->rut = $rut;
       $secretary->dv = $dv;
       $secretary->password = Hash::make(PASSWORD);
-      $secretary->role = 'secretary';
+      //$secretary->role = 'secretary';
       
-      $secretaryInfo->rut = $rut;
-      $secretaryInfo->dv = $dv;
-      $secretaryInfo->idHospital = Input::get('idHospital');
       $secretary->save();
-      $secretaryInfo->save(); 
-      
     }
+    else
+      $secretaryInfo = $secretary->getUserInfoFromHospital(Input::get('idHospital'));
+    
+    if (!isset($secretaryInfo))
+    {
+        $secretaryInfo = new UserInfo();
+        $secretaryInfo->rut = $rut;
+        $secretaryInfo->dv = $dv;
+        $secretaryInfo->idHospital = Input::get('idHospital');
+        $secretaryInfo->save();
+    }
+    
     $secretaryDoctor = new SecretaryDoctor();
     $secretaryDoctor->doctorsRut = Auth::user()->rut;
     $secretaryDoctor->secretarysRut = $rut;
     $secretaryDoctor->idHospital = Input::get('idHospital');
     $secretaryDoctor->save();
     
+    if (!$secretary->isSecretaryOn(Input::get('idHospital')))
+    {
+      $role = new UserRole();
+      $role->rut = $rut;
+      $role->role = "secretary";
+      $role->idHospital = Input::get('idHospital');
+      $role->save();
+    }
+    
+    $secretary["userInfo"] = $secretary->getUserInfoFromHospital(Input::get('idHospital'))->toArray();
     $secretary->pivot = $secretaryDoctor->toArray();
     return $secretary->toJson();
   }
@@ -119,7 +143,7 @@ class DoctorController extends SecretaryController {
       foreach($secretaries as $secretary)
       {
         $completeSecretary = $secretary->toArray();
-        //$completeSecretary['rutFormated'] = $secretary->rutFormated();
+        $completeSecretary['userInfo'] = $secretary->getUserInfoFromHospital($hospital->idHospital)->toArray();
         $completeHospital['secretaries'][] = $completeSecretary;
       }
       $tempReg['hospitals'][] = $completeHospital;
